Updating tests.
