+++ segmentstore/storage/src/test/java/io/pravega/segmentstore/storage/noop/NoOpStorageSystemDataWriteReadTests.java
@@ -15,6 +15,7 @@ import io.pravega.segmentstore.storage.StorageTestBase;
 import io.pravega.segmentstore.storage.SyncStorage;
 import io.pravega.segmentstore.storage.mocks.InMemoryStorageFactory;
 import org.junit.Before;
+import org.junit.Test;
 
 /**
  * This is to test Storage in No-Op mode for system segments only.
@@ -48,4 +49,11 @@ public class NoOpStorageSystemDataWriteReadTests extends StorageTestBase {
     public void testFencing() {
     }
 
+    /**
+     * This method intentionally left blank as it's out of concern for No-Op Storage.
+     * It must be here as it is defined as abstract method in super class.
+     */
+    @Override
+    public void testListSegmentsWithOneSegment() {
+    }
 }
