+++ segmentstore/storage/src/test/java/io/pravega/segmentstore/storage/noop/NoOpStorageUserDataWriteOnlyTests.java
@@ -194,4 +194,7 @@ public class NoOpStorageUserDataWriteOnlyTests extends StorageTestBase {
     public void testFencing() throws Exception {
     }
 
+    @Override
+    public void testListSegmentsWithOneSegment() {
+    }
 }
