+++ segmentstore/storage/src/test/java/io/pravega/segmentstore/storage/noop/NoOpStorageUserDataWriteReadTests.java
@@ -48,4 +48,7 @@ public class NoOpStorageUserDataWriteReadTests extends StorageTestBase {
     public void testFencing() {
     }
 
+    @Override
+    public void testListSegmentsWithOneSegment() {
+    }
 }
